requirements.txt
google-cloud-storage
numpy

